Move interpreters to directory
