+++ generation/src/proof_generation/interpreter/basic_interpreter.py
@@ -4,9 +4,9 @@ from typing import TYPE_CHECKING
 
 from frozendict import frozendict
 
-from proof_generation.aml import App, ESubst, EVar, Exists, Implies, Instantiate, MetaVar, Mu, SSubst, SVar, Symbol, bot
-from proof_generation.interpreter import ExecutionPhase, Interpreter
-from proof_generation.proved import Proved
+from ..aml import App, ESubst, EVar, Exists, Implies, Instantiate, MetaVar, Mu, SSubst, SVar, Symbol, bot
+from ..proved import Proved
+from .interpreter import ExecutionPhase, Interpreter
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
