+++ generation/src/proof_generation/metamath/translate.py
@@ -6,14 +6,12 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 from proof_generation.aml import Implies, MetaVar, Pattern
-from proof_generation.interpreter import ExecutionPhase
-from proof_generation.interpreter_transformer import InterpreterTransformer
+from proof_generation.interpreter import ExecutionPhase, InterpreterTransformer, StatefulInterpreter
 from proof_generation.metamath.converter.converter import MetamathConverter
 from proof_generation.metamath.converter.representation import AxiomWithAntecedents
 from proof_generation.metamath.parser import load_database
 from proof_generation.proof import ProofExp
 from proof_generation.proved import Proved
-from proof_generation.stateful_interpreter import StatefulInterpreter
 
 if TYPE_CHECKING:
     from proof_generation.interpreter import Interpreter
