+++ generation/src/tests/unit/test_instantiation_optimizer.py
@@ -1,7 +1,5 @@
 from proof_generation.aml import Instantiate, MetaVar
-from proof_generation.interpreter import ExecutionPhase
-from proof_generation.optimizing_interpreters import InstantiationOptimizer
-from proof_generation.stateful_interpreter import StatefulInterpreter
+from proof_generation.interpreter import ExecutionPhase, InstantiationOptimizer, StatefulInterpreter
 
 
 def test_instantiation_optimizer() -> None:
